requirements.txt
requests
pyyaml
websocket-client
numpy
geopy
rich
matplotlib
folium
